requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
openai==1.3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-allauth==0.57.0
pytest-django==4.7.0
factory-boy==3.3.0
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0 